backend/requirements.txt
fastapi
uvicorn
akshare
pandas
numpy
requests
pyyaml
apscheduler
